Move getRandomArrayOfNumbers function to puzzle-utils for ease of testing
